requirements.txt
python-ldap
cryptography
sqlalchemy
pymysql
sqlalchemy.orm
ldap3
pycryptodome